requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
langchain-openai>=0.0.5
python-dotenv>=0.19.0
pillow>=10.0.0
requests>=2.31.0 